backend/pyproject.toml
[project]
name = "prison-backend"
version = "0.1.0"
description = "Backend service for prison management"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "passlib[bcrypt]>=1.7.4",
  "PyJWT>=2.9.0",
  "supabase>=2.7.4",
  "alembic>=1.13.0",
  "pyodbc>=5.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "tests*"]
